# api/openapi.yaml
openapi: '3.0.3'
info:
  title: 'Motocorp Dealership Management System API'
  description: |
    A comprehensive cloud-based dealership management system for motorcycle dealerships.
    This API provides multi-tenant architecture with complete CRUD operations for:
    - Inventory Management (motorcycles, parts, accessories)
    - Customer Relationship Management (CRM)
    - Sales Pipeline Management
    - Service Appointment Management
    - User Management with role-based access control
    - Real-time notifications and activity feeds
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`
    
    ## Multi-Tenancy
    The API supports multi-tenant architecture where each organization has its own isolated data.
    The tenant context is automatically resolved from the authenticated user's JWT token.
    
    ## Rate Limiting
    API requests are rate-limited to ensure fair usage and system stability.
    
    ## Error Handling
    Standard HTTP status codes are used. Error responses include detailed error messages
    and validation errors where applicable.
  version: '1.0.0'
  contact:
    name: 'Motocorp DMS Support'
    email: 'support@motocorpdms.com'
    url: 'https://motocorpdms.com/support'
  license:
    name: 'Proprietary'
servers:
  - url: 'https://api.motocorpdms.com/v1'
    description: 'Production server'
...
    description: 'Staging server'
...
security:
  - BearerAuth: []
paths:
...
  /inventory:
    get:
      tags:
        - Inventory
      summary: List inventory items
      description: Retrieves a paginated list of inventory items with filtering and sorting
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: type
          in: query
          description: Filter by inventory type
          schema:
            type: string
            enum: [motorcycle, part, accessory]
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: dealershipId
          in: query
          description: Filter by dealership
          schema:
            type: string
            format: uuid
        - name: lowStock
          in: query
          description: Show only items with low stock
          schema:
            type: boolean
      responses:
        '200':
          description: Inventory items retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInventory'

    post:
      tags:
        - Inventory
      summary: Create inventory item
      description: Creates a new inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInventoryDto'
      responses:
        '201':
          description: Inventory item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /inventory/stats:
    get:
      tags:
        - Inventory
      summary: Get inventory statistics
      description: Retrieves aggregated inventory statistics
      parameters:
        - name: dealershipId
          in: query
          description: Filter statistics by dealership
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Inventory statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryStats'
  /inventory/low-stock:
    get:
      tags:
        - Inventory
      summary: Get low stock items
      description: Retrieves items that are below their reorder point
      parameters:
        - name: dealershipId
          in: query
          description: Filter by dealership
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Low stock items retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItem'
  /inventory/views:
    get:
      tags:
        - Inventory
      summary: List inventory views
      description: Retrieves user's custom inventory views
      responses:
        '200':
          description: Inventory views retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryView'

    post:
      tags:
        - Inventory
      summary: Create inventory view
      description: Creates a new custom inventory view
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInventoryViewDto'
      responses:
        '201':
          description: Inventory view created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryView'
  /inventory/views/{id}:
    get:
      tags:
        - Inventory
      summary: Get inventory view by ID
      description: Retrieves a specific inventory view
      parameters:
        - name: id
          in: path
          required: true
          description: Inventory view ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Inventory view retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryView'

    put:
      tags:
        - Inventory
      summary: Update inventory view
      description: Updates an existing inventory view
      parameters:
        - name: id
          in: path
          required: true
          description: Inventory view ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInventoryViewDto'
      responses:
        '200':
          description: Inventory view updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryView'

    delete:
      tags:
        - Inventory
      summary: Delete inventory view
      description: Deletes an inventory view
      parameters:
        - name: id
          in: path
          required: true
          description: Inventory view ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Inventory view deleted successfully
  /inventory/views/{id}/apply:
    get:
      tags:
        - Inventory
      summary: Apply inventory view
      description: Applies an inventory view with optional additional filters
      parameters:
        - name: id
          in: path
          required: true
          description: Inventory view ID
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Inventory view applied successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInventory'
  /inventory/{id}:
    get:
      tags:
        - Inventory
      summary: Get inventory item by ID
      description: Retrieves a specific inventory item
      parameters:
        - name: id
          in: path
          required: true
          description: Inventory item ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Inventory item retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '404':
          description: Inventory item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Inventory
      summary: Update inventory item
      description: Updates an existing inventory item
      parameters:
        - name: id
          in: path
          required: true
          description: Inventory item ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInventoryDto'
      responses:
        '200':
          description: Inventory item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '404':
          description: Inventory item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Inventory
      summary: Delete inventory item
      description: Soft deletes an inventory item
      parameters:
        - name: id
          in: path
          required: true
          description: Inventory item ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Inventory item deleted successfully
        '404':
          description: Inventory item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /inventory/{id}/stock:
    patch:
      tags:
        - Inventory
      summary: Update inventory stock
      description: Updates the stock quantity of an inventory item
      parameters:
        - name: id
          in: path
          required: true
          description: Inventory item ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockDto'
      responses:
        '200':
          description: Stock updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '404':
          description: Inventory item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Sales Management
...
  /notifications:
    get:
      tags:
        - Notifications
      summary: List notifications
      description: Retrieves user's notifications
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: type
          in: query
          description: Filter by notification type
          schema:
            type: string
            enum: [inventory_alert, sales_update, service_reminder]
        - name: unread
          in: query
          description: Show only unread notifications
          schema:
            type: boolean
      responses:
        '200':
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNotifications'

    post:
      tags:
        - Notifications
      summary: Create notification
      description: Creates a new notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotificationDto'
      responses:
        '201':
          description: Notification created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
  /notifications/{id}:
    get:
      tags:
        - Notifications
      summary: Get notification by ID
      description: Retrieves a specific notification
      parameters:
        - name: id
          in: path
          required: true
          description: Notification ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'

    put:
      tags:
        - Notifications
      summary: Update notification
      description: Updates a notification (e.g., mark as read)
      parameters:
        - name: id
          in: path
          required: true
          description: Notification ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNotificationDto'
      responses:
        '200':
          description: Notification updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'

    delete:
      tags:
        - Notifications
      summary: Delete notification
      description: Deletes a notification
      parameters:
        - name: id
          in: path
          required: true
          description: Notification ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Notification deleted successfully

  # Activity Feed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Bearer token authentication. Format: Bearer <your_jwt_token>'
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination (starts from 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    InventoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique inventory item identifier
          example: '123e4567-e89b-12d3-a456-426614174000'
        dealershipId:
          type: string
          format: uuid
          description: Dealership ID
          example: '123e4567-e89b-12d3-a456-426614174001'
        sku:
          type: string
          description: Stock Keeping Unit
          example: 'MOT-001-2024'
        type:
          type: string
          enum: [motorcycle, part, accessory]
          description: Type of inventory item
          example: 'motorcycle'
        category:
          type: string
          nullable: true
          description: Item category
          example: 'Sport'
        name:
          type: string
          description: Item name
          example: 'Yamaha YZF-R1'
        description:
          type: string
          nullable: true
          description: Item description
          example: 'High-performance sport motorcycle'
        costPrice:
          type: number
          format: decimal
          nullable: true
          description: Cost price
          example: 12500.00
        sellingPrice:
          type: number
          format: decimal
          description: Selling price
          example: 16999.99
        quantityOnHand:
          type: integer
          description: Current stock quantity
          example: 3
        reorderPoint:
          type: integer
          nullable: true
          description: Reorder point for low stock alerts
          example: 2
        supplierInfo:
          type: object
          nullable: true
          properties:
            name:
              type: string
              example: 'Yamaha Motor Corp'
            contactEmail:
              type: string
              format: email
              example: 'orders@yamaha.com'
            contactPhone:
              type: string
              example: '+1-800-962-7926'
            address:
              type: string
              example: '6555 Katella Ave, Cypress, CA 90630'
            leadTime:
              type: integer
              description: Lead time in days
              example: 14
            minimumOrder:
              type: integer
              description: Minimum order quantity
              example: 1
        isActive:
          type: boolean
          description: Whether the item is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2024-01-15T10:30:00Z'
    CreateInventoryDto:
      type: object
      required:
        - dealershipId
        - sku
        - type
        - name
        - sellingPrice
      properties:
        dealershipId:
          type: string
          format: uuid
          description: Dealership ID
          example: '123e4567-e89b-12d3-a456-426614174001'
        sku:
          type: string
          description: Stock Keeping Unit
          example: 'MOT-002-2024'
        type:
          type: string
          enum: [motorcycle, part, accessory]
          description: Type of inventory item
          example: 'motorcycle'
        category:
          type: string
          nullable: true
          description: Item category
          example: 'Cruiser'
        name:
          type: string
          description: Item name
          example: 'Honda Rebel 300'
        description:
          type: string
          nullable: true
          description: Item description
          example: 'Entry-level cruiser motorcycle'
        costPrice:
          type: number
          format: decimal
          nullable: true
          description: Cost price
          example: 3500.00
        sellingPrice:
          type: number
          format: decimal
          description: Selling price
          example: 4699.99
        quantityOnHand:
          type: integer
          minimum: 0
          description: Initial stock quantity
          example: 5
          default: 0
        reorderPoint:
          type: integer
          minimum: 0
          nullable: true
          description: Reorder point for low stock alerts
          example: 1
        supplierInfo:
          type: object
          nullable: true
          properties:
            name:
              type: string
              example: 'Honda Motor Co'
            contactEmail:
              type: string
              format: email
              example: 'orders@honda.com'
            contactPhone:
              type: string
              example: '+1-800-999-1009'
            address:
              type: string
              example: '1919 Torrance Blvd, Torrance, CA 90501'
            leadTime:
              type: integer
              description: Lead time in days
              example: 10
            minimumOrder:
              type: integer
              description: Minimum order quantity
              example: 1
...
    UpdateStockDto:
      type: object
      required:
        - operation
        - quantity
      properties:
        operation:
          type: string
          enum: [add, subtract, set]
          description: Stock operation type
          example: 'add'
        quantity:
          type: integer
          minimum: 0
          description: Quantity for the operation
          example: 5
        reason:
          type: string
          nullable: true
          description: Reason for stock update
          example: 'New shipment received'
    InventoryView:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique view identifier
          example: '123e4567-e89b-12d3-a456-426614174000'
        userId:
          type: string
          format: uuid
          description: User ID who created the view
          example: '123e4567-e89b-12d3-a456-426614174001'
        dealershipId:
          type: string
          format: uuid
          nullable: true
          description: Dealership ID (if view is dealership-specific)
          example: '123e4567-e89b-12d3-a456-426614174002'
        name:
          type: string
          description: View name
          example: 'Low Stock Motorcycles'
        description:
          type: string
          nullable: true
          description: View description
          example: 'Motorcycles with stock below reorder point'
        isDefault:
          type: boolean
          description: Whether this is the default view
          example: false
        isShared:
          type: boolean
          description: Whether this view is shared with other users
          example: true
        filterConfig:
          type: object
          description: Filter configuration
          example: {"type": "motorcycle", "quantityOnHand": {"$lt": "reorderPoint"}}
        sortConfig:
          type: object
          description: Sort configuration
          example: {"field": "quantityOnHand", "direction": "asc"}
        groupConfig:
          type: object
          nullable: true
          description: Group configuration
          example: {"field": "category"}
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2024-01-15T10:30:00Z'
    CreateInventoryViewDto:
      type: object
      required:
        - name
        - filterConfig
        - sortConfig
      properties:
        name:
          type: string
          description: View name
          example: 'High Value Items'
        description:
          type: string
          nullable: true
          description: View description
          example: 'Items with selling price above $10,000'
        dealershipId:
          type: string
          format: uuid
          nullable: true
          description: Dealership ID (if view is dealership-specific)
          example: '123e4567-e89b-12d3-a456-426614174002'
        isDefault:
          type: boolean
          description: Whether this is the default view
          example: false
          default: false
        isShared:
          type: boolean
          description: Whether this view is shared with other users
          example: false
          default: false
        filterConfig:
          type: object
          description: Filter configuration
          example: {"sellingPrice": {"$gte": 10000}}
        sortConfig:
          type: object
          description: Sort configuration
          example: {"field": "sellingPrice", "direction": "desc"}
        groupConfig:
          type: object
          nullable: true
          description: Group configuration
          example: {"field": "type"}
    UpdateInventoryViewDto:
      type: object
      properties:
        name:
          type: string
          description: View name
          example: 'Premium Items'
        description:
          type: string
          nullable: true
          description: View description
          example: 'Premium motorcycles and accessories'
        isDefault:
          type: boolean
          description: Whether this is the default view
          example: true
        isShared:
          type: boolean
          description: Whether this view is shared with other users
          example: true
        filterConfig:
          type: object
          description: Filter configuration
          example: {"sellingPrice": {"$gte": 15000}}
        sortConfig:
          type: object
          description: Sort configuration
          example: {"field": "name", "direction": "asc"}
        groupConfig:
          type: object
          nullable: true
          description: Group configuration
          example: {"field": "category"}
    PaginatedInventory:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/InventoryItem'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    InventoryStats:
      type: object
      properties:
        totalItems:
          type: integer
          description: Total number of inventory items
          example: 250
        totalValue:
          type: number
          format: decimal
          description: Total inventory value (selling price)
          example: 1250000.00
        lowStockItems:
          type: integer
          description: Number of items below reorder point
          example: 15
        outOfStockItems:
          type: integer
          description: Number of items with zero stock
          example: 3
        byType:
          type: object
          properties:
            motorcycle:
              type: integer
              example: 45
            part:
              type: integer
              example: 180
            accessory:
              type: integer
              example: 25
        byCategory:
          type: object
          additionalProperties:
            type: integer
          example:
            Sport: 15
            Cruiser: 12
            Touring: 8
            Adventure: 10

    # Sales Management Schemas
...
    CreateNotificationDto:
      type: object
      required:
        - userId
        - type
        - title
        - message
      properties:
        userId:
          type: string
          format: uuid
          description: User ID
          example: '123e4567-e89b-12d3-a456-426614174001'
        type:
          type: string
          enum: [inventory_alert, sales_update, service_reminder]
          description: Notification type
          example: 'sales_update'
        title:
          type: string
          description: Notification title
          example: 'Sale Stage Updated'
        message:
          type: string
          description: Notification message
          example: 'Sale #12345 moved to negotiation stage'
        data:
          type: object
          nullable: true
          description: Additional notification data
          example: {"saleId": "123e4567-e89b-12d3-a456-426614174003", "newStage": "negotiation"}
...
    PaginatedNotifications:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'

    # Activity Feed Schemas
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Human readable error message
          example: 'Resource not found'
        error:
          type: string
          description: Error type
          example: 'NotFound'
        statusCode:
          type: integer
          description: HTTP status code
          example: 404
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2024-01-15T10:30:00Z'
        path:
          type: string
          description: Request path that caused the error
          example: '/api/v1/customers/invalid-id'
    ValidationError:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
          description: General validation error message
          example: 'Validation failed'
        error:
          type: string
          description: Error type
          example: 'Bad Request'
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2024-01-15T10:30:00Z'
        path:
          type: string
          description: Request path that caused the error
          example: '/api/v1/customers'
        errors:
          type: array
          description: Detailed validation errors
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that failed validation
                example: 'email'
              message:
                type: string
                description: Validation error message
                example: 'must be a valid email address'
              value:
                description: Invalid value that was provided
                example: 'invalid-email'
...